+++ pkg/stack/stack.go
@@ -43,18 +43,17 @@ type CpuMemory struct {
 	Memory string `yaml:"memory,omitempty"`
 }
 
-// FunctionResources Memory and CPU
-//type FunctionResources struct {
-//	Memory string `yaml:"memory"`
-//	CPU    string `yaml:"cpu"`
-//}
-
 // Stack root level YAML file to define FaaS function-set
 type Stack struct {
-	Version     string              `yaml:"version,omitempty"`
-	Provider    Provider            `yaml:"provider,omitempty"`
-	Hostname    string              `yaml:"hostname,omitempty"`
-	Functions   map[string]Function `yaml:"functions,omitempty"`
-	StackConfig Config              `yaml:"config,omitempty"`
-	Custom      []string            `yaml:"custom,omitempty"`
+	Version            string              `yaml:"version,omitempty"`
+	Provider           Provider            `yaml:"provider,omitempty"`
+	Hostname           string              `yaml:"hostname,omitempty"`
+	Functions          map[string]Function `yaml:"functions,omitempty"`
+	StackConfig        Config              `yaml:"config,omitempty"`
+	Custom             []string            `yaml:"custom,omitempty"`
+	KongAssetsEnabled  bool                `yaml:"kongAssetsEnabled,omitempty"`
+}
+
+func GetYamlFileName() string {
+	return stackFileName
 }