+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release Notes
 
+## Version v0.0.6 - 2020-07-00
+
+Features:
+
+- Added new flag `kubeconfig` in `safira function remove`
+- Added new command `safira kongplugin`
+- Added new subcommand `safira kongplugin new`
+
+Improvements:
+
+- The link to the swagger editor has been added to the output of `safira infra up`
+- The removed suffix from deploys at output of `safira infra status`
+- Added output of URLs of the `safira infra status` command
+
 ## Version v0.0.5 - 2020-06-30
 
 Features: